Moved standard draw functions to Graphics System/General Redundant.
